refactor: Refactor directory structure
