--- src/Register.tsx
@@ -1,7 +1,0 @@
-import SignUp from './copied-material/sign-up/SignUp.tsx';
-
-function Login() {
-  return <SignUp />;
-}
-
-export default Login;
