+++ src/components/home.tsx
@@ -1,7 +1,5 @@
 import AppContainer from './app-container.tsx';
 
-function Home() {
+export default function Home() {
   return <AppContainer title='This is Home'>Home</AppContainer>;
 }
-
-export default Home;
