+++ src/components/login.tsx
@@ -1,4 +1,4 @@
-import SignIn from './copied-material/sign-in/SignIn.tsx';
+import SignIn from '../copied-material/sign-in/SignIn.tsx';
 
 function Login() {
   return <SignIn />;
