+++ src/components/preferences.tsx
@@ -1,0 +1,3 @@
+export default function Preferences() {
+  return 'Hello from preferences';
+}
