+++ src/components/register.tsx
@@ -1,0 +1,5 @@
+import SignUp from '../copied-material/sign-up/SignUp.tsx';
+
+export default function Login() {
+  return <SignUp />;
+}
