--- src/utils/authorized-fetch.ts
@@ -1,13 +1,0 @@
-export function authorizedFetch(
-  url: string,
-
-  options?: RequestInit
-) {
-  return fetch(url, {
-    ...options,
-    headers: {
-      ...options?.headers,
-      'Content-Type': 'application/json',
-    },
-  });
-}
